requirements.txt
backcall==0.1.0
bleach==3.0.2
cloudpickle==0.6.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
dicom==0.9.9.post1
entrypoints==0.2.3
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkg-resources==0.0.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pydicom==1.2.1
Pygments==2.3.0
pyparsing==2.3.0
PySpaceNav==0.1
python-dateutil==2.7.5
pyzmq==17.1.2
qtconsole==4.4.3
scikit-learn==0.20.2
scipy==1.2.0
Send2Trash==1.5.0
six==1.11.0
sklearn==0.0
spnav==0.9
sympy==1.3
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
transforms3d==0.3.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
